Create serializer module and reuse it across the application
