+++ lib/paper_trail.ex
@@ -3,6 +3,7 @@ defmodule PaperTrail do
 
   alias PaperTrail.Version
   alias PaperTrail.RepoClient
+  alias PaperTrail.Serializer
 
   defdelegate get_version(record), to: PaperTrail.VersionQueries
   defdelegate get_version(model_or_record, id_or_options), to: PaperTrail.VersionQueries
@@ -11,6 +12,13 @@ defmodule PaperTrail do
   defdelegate get_versions(model_or_record, id_or_options), to: PaperTrail.VersionQueries
   defdelegate get_versions(model, id, options), to: PaperTrail.VersionQueries
   defdelegate get_current_model(version), to: PaperTrail.VersionQueries
+  defdelegate make_version_struct(version, model, options), to: Serializer
+  defdelegate get_sequence_from_model(changeset), to: Serializer
+  defdelegate serialize(data), to: Serializer
+  defdelegate get_sequence_id(table_name), to: Serializer
+  defdelegate add_prefix(changeset, prefix), to: Serializer
+  defdelegate get_item_type(data), to: Serializer
+  defdelegate get_model_id(model), to: Serializer
 
   @doc """
   Inserts a record to the database with a related version insertion in one transaction
@@ -138,108 +146,5 @@ defmodule PaperTrail do
       model
     end)
     |> elem(1)
-  end
-
-  defp make_version_struct(%{event: "insert"}, model, options) do
-    originator = PaperTrail.RepoClient.originator()
-    originator_ref = options[originator[:name]] || options[:originator]
-
-    %Version{
-      event: "insert",
-      item_type: get_item_type(model),
-      item_id: get_model_id(model),
-      item_changes: serialize(model),
-      originator_id:
-        case originator_ref do
-          nil -> nil
-          _ -> originator_ref |> Map.get(:id)
-        end,
-      origin: options[:origin],
-      meta: options[:meta]
-    }
-    |> add_prefix(options[:prefix])
-  end
-
-  defp make_version_struct(%{event: "update"}, changeset, options) do
-    originator = PaperTrail.RepoClient.originator()
-    originator_ref = options[originator[:name]] || options[:originator]
-
-    %Version{
-      event: "update",
-      item_type: get_item_type(changeset),
-      item_id: get_model_id(changeset),
-      item_changes: changeset.changes,
-      originator_id:
-        case originator_ref do
-          nil -> nil
-          _ -> originator_ref |> Map.get(:id)
-        end,
-      origin: options[:origin],
-      meta: options[:meta]
-    }
-    |> add_prefix(options[:prefix])
-  end
-
-  defp make_version_struct(%{event: "delete"}, model_or_changeset, options) do
-    originator = PaperTrail.RepoClient.originator()
-    originator_ref = options[originator[:name]] || options[:originator]
-
-    %Version{
-      event: "delete",
-      item_type: get_item_type(model_or_changeset),
-      item_id: get_model_id(model_or_changeset),
-      item_changes: serialize(model_or_changeset),
-      originator_id:
-        case originator_ref do
-          nil -> nil
-          _ -> originator_ref |> Map.get(:id)
-        end,
-      origin: options[:origin],
-      meta: options[:meta]
-    }
-    |> add_prefix(options[:prefix])
-  end
-
-  defp get_sequence_from_model(changeset) do
-    table_name =
-      case Map.get(changeset, :data) do
-        nil -> changeset.__struct__.__schema__(:source)
-        _ -> changeset.data.__struct__.__schema__(:source)
-      end
-
-    get_sequence_id(table_name)
-  end
-
-  defp get_sequence_id(table_name) do
-    Ecto.Adapters.SQL.query!(RepoClient.repo(), "select last_value FROM #{table_name}_id_seq").rows
-    |> List.first()
-    |> List.first()
-  end
-
-  defp serialize(%Ecto.Changeset{data: data}), do: serialize(data)
-
-  defp serialize(model) do
-    relationships = model.__struct__.__schema__(:associations)
-    Map.drop(model, [:__struct__, :__meta__] ++ relationships)
-  end
-
-  defp add_prefix(changeset, nil), do: changeset
-  defp add_prefix(changeset, prefix), do: Ecto.put_meta(changeset, prefix: prefix)
-
-  defp get_item_type(%Ecto.Changeset{data: data}), do: get_item_type(data)
-  defp get_item_type(model), do: model.__struct__ |> Module.split() |> List.last()
-
-  def get_model_id(%Ecto.Changeset{data: data}), do: get_model_id(data)
-
-  def get_model_id(model) do
-    {_, model_id} = List.first(Ecto.primary_key(model))
-
-    case PaperTrail.Version.__schema__(:type, :item_id) do
-      :integer ->
-        model_id
-
-      _ ->
-        "#{model_id}"
-    end
   end
 end
