+++ lib/paper_trail/multi.ex
@@ -4,6 +4,7 @@ defmodule PaperTrail.Multi do
   alias PaperTrail
   alias PaperTrail.Version
   alias PaperTrail.RepoClient
+  alias PaperTrail.Serializer
 
   defdelegate new(), to: Ecto.Multi
   defdelegate append(lhs, rhs), to: Ecto.Multi
@@ -14,6 +15,13 @@ defmodule PaperTrail.Multi do
   defdelegate run(multi, name, run), to: Ecto.Multi
   defdelegate run(multi, name, mod, fun, args), to: Ecto.Multi
   defdelegate to_list(multi), to: Ecto.Multi
+  defdelegate make_version_struct(version, model, options), to: Serializer
+  defdelegate get_sequence_from_model(changeset), to: Serializer
+  defdelegate serialize(data), to: Serializer
+  defdelegate get_sequence_id(table_name), to: Serializer
+  defdelegate add_prefix(changeset, prefix), to: Serializer
+  defdelegate get_item_type(data), to: Serializer
+  defdelegate get_model_id(model), to: Serializer
 
   def insert(%Ecto.Multi{} = multi, changeset, options \\ [
     origin: nil, meta: nil, originator: nil, prefix: nil, model_key: :model, version_key: :version, ecto_options: []
@@ -144,109 +152,6 @@ defmodule PaperTrail.Multi do
           {:error, :model, changeset, %{}} -> {:error, Map.merge(changeset, %{repo: repo})}
           _ -> transaction
         end
-    end
-  end
-
-  defp make_version_struct(%{event: "insert"}, model, options) do
-    originator = PaperTrail.RepoClient.originator()
-    originator_ref = options[originator[:name]] || options[:originator]
-
-    %Version{
-      event: "insert",
-      item_type: get_item_type(model),
-      item_id: get_model_id(model),
-      item_changes: serialize(model),
-      originator_id:
-        case originator_ref do
-          nil -> nil
-          _ -> originator_ref |> Map.get(:id)
-        end,
-      origin: options[:origin],
-      meta: options[:meta]
-    }
-    |> add_prefix(options[:prefix])
-  end
-
-  defp make_version_struct(%{event: "update"}, changeset, options) do
-    originator = PaperTrail.RepoClient.originator()
-    originator_ref = options[originator[:name]] || options[:originator]
-
-    %Version{
-      event: "update",
-      item_type: get_item_type(changeset),
-      item_id: get_model_id(changeset),
-      item_changes: changeset.changes,
-      originator_id:
-        case originator_ref do
-          nil -> nil
-          _ -> originator_ref |> Map.get(:id)
-        end,
-      origin: options[:origin],
-      meta: options[:meta]
-    }
-    |> add_prefix(options[:prefix])
-  end
-
-  defp make_version_struct(%{event: "delete"}, model_or_changeset, options) do
-    originator = PaperTrail.RepoClient.originator()
-    originator_ref = options[originator[:name]] || options[:originator]
-
-    %Version{
-      event: "delete",
-      item_type: get_item_type(model_or_changeset),
-      item_id: get_model_id(model_or_changeset),
-      item_changes: serialize(model_or_changeset),
-      originator_id:
-        case originator_ref do
-          nil -> nil
-          _ -> originator_ref |> Map.get(:id)
-        end,
-      origin: options[:origin],
-      meta: options[:meta]
-    }
-    |> add_prefix(options[:prefix])
-  end
-
-  defp get_sequence_from_model(changeset) do
-    table_name =
-      case Map.get(changeset, :data) do
-        nil -> changeset.__struct__.__schema__(:source)
-        _ -> changeset.data.__struct__.__schema__(:source)
-      end
-
-    get_sequence_id(table_name)
-  end
-
-  defp get_sequence_id(table_name) do
-    Ecto.Adapters.SQL.query!(RepoClient.repo(), "select last_value FROM #{table_name}_id_seq").rows
-    |> List.first()
-    |> List.first()
-  end
-
-  defp serialize(%Ecto.Changeset{data: data}), do: serialize(data)
-
-  defp serialize(model) do
-    relationships = model.__struct__.__schema__(:associations)
-    Map.drop(model, [:__struct__, :__meta__] ++ relationships)
-  end
-
-  defp add_prefix(changeset, nil), do: changeset
-  defp add_prefix(changeset, prefix), do: Ecto.put_meta(changeset, prefix: prefix)
-
-  defp get_item_type(%Ecto.Changeset{data: data}), do: get_item_type(data)
-  defp get_item_type(model), do: model.__struct__ |> Module.split() |> List.last()
-
-  defp get_model_id(%Ecto.Changeset{data: data}), do: get_model_id(data)
-
-  defp get_model_id(model) do
-    {_, model_id} = List.first(Ecto.primary_key(model))
-
-    case PaperTrail.Version.__schema__(:type, :item_id) do
-      :integer ->
-        model_id
-
-      _ ->
-        "#{model_id}"
     end
   end
 end
