+++ test/paper_trail/bang_functions_strict_mode_test.exs
@@ -7,6 +7,8 @@ defmodule PaperTrailTest.StrictModeBangFunctions do
   alias StrictCompany, as: Company
   alias StrictPerson, as: Person
   alias PaperTrailTest.MultiTenantHelper, as: MultiTenant
+  alias PaperTrail.RepoClient
+  alias PaperTrail.Serializer
 
   @create_company_params %{name: "Acme LLC", is_active: true, city: "Greenwich"}
   @update_company_params %{
@@ -14,6 +16,9 @@ defmodule PaperTrailTest.StrictModeBangFunctions do
     website: "http://www.acme.com",
     facebook: "acme.llc"
   }
+
+  defdelegate repo, to: RepoClient
+  defdelegate serialize(data), to: Serializer
 
   doctest PaperTrail
 
@@ -1020,11 +1025,6 @@ defmodule PaperTrailTest.StrictModeBangFunctions do
     first(Person, :id) |> MultiTenant.add_prefix_to_query() |> repo().one()
   end
 
-  defp serialize(model) do
-    relationships = model.__struct__.__schema__(:associations)
-    Map.drop(model, [:__struct__, :__meta__] ++ relationships)
-  end
-
   defp reset_all_data() do
     repo().delete_all(Person)
     repo().delete_all(Company)
@@ -1047,7 +1047,4 @@ defmodule PaperTrailTest.StrictModeBangFunctions do
     map |> Jason.encode!() |> Jason.decode!()
   end
 
-  defp repo() do
-    PaperTrail.RepoClient.repo()
-  end
 end
