+++ test/paper_trail/strict_mode_test.exs
@@ -7,6 +7,7 @@ defmodule PaperTrailStrictModeTest do
   alias PaperTrail.Version
   alias StrictCompany, as: Company
   alias StrictPerson, as: Person
+  alias PaperTrail.Serializer
 
   @repo PaperTrail.RepoClient.repo()
   @create_company_params %{name: "Acme LLC", is_active: true, city: "Greenwich"}
@@ -15,6 +16,8 @@ defmodule PaperTrailStrictModeTest do
     website: "http://www.acme.com",
     facebook: "acme.llc"
   }
+
+  defdelegate serialize(data), to: Serializer
 
   doctest PaperTrail
 
@@ -498,10 +501,5 @@ defmodule PaperTrailStrictModeTest do
 
   defp update_company_with_version(company, params \\ @update_company_params, options \\ []) do
     Company.changeset(company, params) |> PaperTrail.update(options)
-  end
-
-  defp serialize(model) do
-    relationships = model.__struct__.__schema__(:associations)
-    Map.drop(model, [:__struct__, :__meta__] ++ relationships)
   end
 end
