+++ test/version/version_queries_test.exs
@@ -5,8 +5,11 @@ defmodule PaperTrailTest.VersionQueries do
   alias SimpleCompany, as: Company
   alias SimplePerson, as: Person
   alias PaperTrailTest.MultiTenantHelper, as: MultiTenant
+  alias PaperTrail.RepoClient
 
   import Ecto.Query
+
+  defdelegate repo, to: RepoClient
 
   setup_all do
     Application.put_env(:paper_trail, :repo, PaperTrail.Repo)
@@ -191,9 +194,5 @@ defmodule PaperTrailTest.VersionQueries do
     Version
     |> MultiTenant.add_prefix_to_query()
     |> repo().delete_all()
-  end
-
-  defp repo() do
-    PaperTrail.RepoClient.repo()
   end
 end
